+++ core/docs/changelog.md
@@ -1,6 +1,7 @@
 # TBD
 ### Features
 * The API container will now shut down all the containers in its network as it shuts down, which is a step towards enclaves being independent of the testing framework
+* Created `EnclaveManager`, to start & stop enclaves independent of the testing framework
 
 # 1.18.0
 ### Fixes