+++ core/docs/changelog.md
@@ -1,4 +1,7 @@
 # TBD
+### Fixes
+* Correct links now that `kurtosis-libs` is renamed to `kurtosis-testsuite-starter-pack`
+
 ### Features
 * The API container will now shut down all the containers in its network as it shuts down, which is a step towards enclaves being independent of the testing framework
 * Created `EnclaveManager`, to start & stop enclaves independent of the testing framework