+++ core/docs/changelog.md
@@ -1,6 +1,7 @@
 # TBD
 ### Fixes
 * Correct links now that `kurtosis-libs` is renamed to `kurtosis-testsuite-starter-pack`
+* There are more disallowed IP ranges than just the multicast addresses (see [this Wikipedia article](https://en.wikipedia.org/wiki/IPv4#Special-use_addresses)), so prevent the Docker network allocator from choosing those
 
 ### Features
 * The API container will now shut down all the containers in its network as it shuts down, which is a step towards enclaves being independent of the testing framework